% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intergrowth_constants.R
\name{estimate_sga_intergrowth}
\alias{estimate_sga_intergrowth}
\title{Return SGA given a dataframe of sex, weight, and gestational age at birth}
\usage{
estimate_sga_intergrowth(df, include_outliers = FALSE, growth_standard = igs())
}
\arguments{
\item{df}{The dataframe containing the three required variables, as above}

\item{include_outliers}{(logical, default = FALSE); set to true if individuals
with non-missing gestational age value below the minimum value (24) or above
the maximum value (default 43) should be assumed to have gestage set to those
bounds}

\item{growth_standard}{This is a growth standard as produced by the \code{igs()} function
The default value is igs() (i.e. without any parameters), which is the intergrowth
standard}
}
\value{
factor vector
}
\description{
Function will return an integer vector of class \code{factor} holding values
AGA (0), SGA 3-10\% (1) or SGA <3\% (2), given a data frame / tibble/ or
data.table that has (at least) columns named 'gestage','weight',and 'sex'.
}
\details{
There are some requirements: (1) the 'gestage' column must not be negative, and will
be assumed to be expressed in weeks (fractional weeks are allowed); (2) the
'sex' column can only contain values 1,2 or NA. Values equal to 1 will be
assumed to indicate males, while values equal to 2 will be assumed to indicate
females; (3) all three required variables must be numeric; (4) the 'weight'
variable will be assumed to be in grams. Multiply by 1000 if your weights
are in kilograms.
}
\examples{
estimate_sga_intergrowth(data)
}
