% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intergrowth_constants.R
\name{fix_outliers}
\alias{fix_outliers}
\title{Function to change outliers so they equal the endpoints of a growth standard}
\usage{
fix_outliers(df, gs)
}
\description{
Function to change outliers so they equal the endpoints of a growth standard
}
\keyword{internal}
