% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intergrowth_constants.R
\name{gest_age_cutpoints}
\alias{gest_age_cutpoints}
\title{Function returns a list of cutpoints, given a min, max, and length out}
\usage{
gest_age_cutpoints(min = 24, max = 43, length = 134)
}
\arguments{
\item{min}{starting point for gestational age cutpoint (default = 24)}

\item{max}{ending point for gestational age cutpoints (default = 43)}

\item{length}{number of cutpoints (default = 134)}
}
\description{
Function returns a list of cutpoints, given a min, max, and length out
}
\keyword{internal}
