% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intergrowth_constants.R
\name{igs}
\alias{igs}
\title{Function returns a data table of intergrowth standards}
\usage{
igs(growth_constants = NULL, min = 24, max = 43, length = 134)
}
\arguments{
\item{growth_constants}{(default is NULL); this should be a function that
returns a names list of weights. The named list must be length=4, and the
names must be "boys_10","boys_3","girls_10", and "girls_3". Each element
of the list must be vector of weights in grams, and each should correspond
to a gestational age cutpoints. The cutpoints must be equally spaced sequence
that is generated by passing in min, max, and length.}

\item{min}{starting point for gestational age cutpoint (default = 24)}

\item{max}{ending point for gestational age cutpoints (default = 43)}

\item{length}{number of cutpoints (default = 134)}
}
\description{
This function will return a data.table of intergrowth standards, with
column g_start and g_end indicating the start and end of a gestational
age range, and columns boys_10, boys_3, girls_10, and girls_3 indicating
the weight at which infants falling within that range must meet or exceed
in order to avoid being classified as SGA.
}
\details{
The function will use intergrowth standards by default; however a custom
set of standards can be generate by passing in parameters as below
}
