% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intergrowth_functions.R
\name{intergrowth_constants}
\alias{intergrowth_constants}
\title{Function returns a list of Intergrowth-21st standards
https://intergrowth21.tghn.org/standards-tools/}
\usage{
intergrowth_constants()
}
\description{
Function returns a list of Intergrowth-21st standards
https://intergrowth21.tghn.org/standards-tools/
}
\keyword{internal}
