% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intergrowth_constants.R
\name{intergrowth_constants}
\alias{intergrowth_constants}
\title{Function returns a list of intergrowth standards}
\usage{
intergrowth_constants()
}
\description{
Function returns a list of intergrowth standards
}
\keyword{internal}
